requirements.production.txt
# Core dependencies
numpy==1.24.3
pandas==1.5.3
geopandas==0.13.2
folium==0.15.1
branca==0.7.1

# Web frameworks
dash==2.9.3
dash-bootstrap-components==1.4.1
fastapi==0.95.0
Flask==2.2.3
Werkzeug==2.2.3

# Production server
gunicorn==20.1.0
uvicorn==0.21.1

# Streamlit for ML apps
streamlit==1.28.1

# Plotting and visualization
plotly==5.13.1
matplotlib==3.7.1
seaborn==0.12.2

# Database
psycopg2-binary==2.9.5
SQLAlchemy==2.0.7

# Environment and config
python-dotenv==1.0.0

# ML and data processing
scikit-learn==1.3.0
joblib==1.3.2

# Caching and sessions
redis==4.6.0
Flask-Session==0.5.0

# Security
flask-talisman==1.1.0
cryptography==41.0.3

# Monitoring and logging
prometheus-client==0.17.1
structlog==23.1.0

# Utils
requests==2.31.0
python-dateutil==2.8.2 